:recycle: Extract auth_accessor.go from auth.go
